почти переделал tsv-file-reader под свою задачу, допилил Offering
