+++ mocks/offerings.tsv
@@ -1,2 +1,2 @@
-Королевские палаты в пяти минутах от Елисейских полей	2022-09-10T08:45:40.284Z	https://media.architecturaldigest.com/photos/5e6a42b25c94700009daa7d7/16:9/w_2580%2Cc_limit/AD0420_DIRAND_7.jpg	[https://media.architecturaldigest.com/photos/5e6a42b25c94700009daa7d7/16:9/w_2580%2Cc_limit/AD0420_DIRAND_7.jpg, https://media.architecturaldigest.com/photos/61e04ed6f3d6076b395bf303/master/w_2580%2Cc_limit/G19_4416.jpg, https://media.architecturaldigest.com/photos/5a4d0b5fcf0d6c31eb82efcd/master/w_2580%2Cc_limit/REDA%2520AMALOU%2520DESIGN%2520APARTMENT%2520%25C2%25A9%2520FRANCIS%2520AMIAND%2520(2).jpg, https://media.architecturaldigest.com/photos/62d6d05f0e13d40b82bb49cf/master/w_2580%2Cc_limit/G19_4619%2520copy.jpg, https://media.architecturaldigest.com/photos/6377c1d2abb4b34435a704dc/master/w_2580%2Cc_limit/0113-AD-DENI-06%2520copy.jpg, https://media.architecturaldigest.com/photos/6377c1819bc479d85214d271/master/w_2580%2Cc_limit/1213-AD-GARC-05%2520copy.jpg]	true	false	5	apartment	7	20	3000	[Breakfast, Towels, Fridge]	Jean-Batist
-Роскошная двушка в историческом районе Дюссельдорфа	2022-10-01T08:45:40.284Z	https://cf.bstatic.com/xdata/images/hotel/max1280x900/326256908.jpg?k=bdd20df0174cd044f2ec5d975c3586d019b90387a4ea7a2bdca37d42965c1337&o=&hp=1	[https://cf.bstatic.com/xdata/images/hotel/max1280x900/326256908.jpg?k=bdd20df0174cd044f2ec5d975c3586d019b90387a4ea7a2bdca37d42965c1337&o=&hp=1, https://cf.bstatic.com/xdata/images/hotel/max1280x900/326256972.jpg?k=890fc5ba61c48d55487cf087ab0b906e8193250a2567eb4649bb865c9b425c23&o=&hp=1, https://cf.bstatic.com/xdata/images/hotel/max1280x900/326257199.jpg?k=0b60c5afecbafcc358a5fa0369e30b0c6d95c2a76ecbe614cb3157d98fcb3b34&o=&hp=1, https://cf.bstatic.com/xdata/images/hotel/max1280x900/326256930.jpg?k=e7d575f736455a64f81144f4b46a70e0061f0a31a15f415fc22e1922a11ec14a&o=&hp=1, https://cf.bstatic.com/xdata/images/hotel/max1280x900/326256937.jpg?k=cd24fd744c417b25a06e5b3b5f5b32f4a5124321c651ff420efd1c997958db3e&o=&hp=1, https://cf.bstatic.com/xdata/images/hotel/max1280x900/326256936.jpg?k=e8643617e246b762df649d5af59685f3b7c447d3e2692974302a0139cd1c77a7&o=&hp=1]	true	true	5	apartment	2	2	1000	[Breakfast, Towels, Fridge, Air conditioning, Laptop friendly workspace, Baby seat]	Hanz
+Королевские палаты в пяти минутах от Елисейских полей	2022-09-10T08:45:40.284Z	https://media.architecturaldigest.com/photos/5e6a42b25c94700009daa7d7/16:9/w_2580%2Cc_limit/AD0420_DIRAND_7.jpg	https://media.architecturaldigest.com/photos/5e6a42b25c94700009daa7d7/16:9/w_2580%2Cc_limit/AD0420_DIRAND_7.jpg;https://media.architecturaldigest.com/photos/61e04ed6f3d6076b395bf303/master/w_2580%2Cc_limit/G19_4416.jpg;https://media.architecturaldigest.com/photos/5a4d0b5fcf0d6c31eb82efcd/master/w_2580%2Cc_limit/REDA%2520AMALOU%2520DESIGN%2520APARTMENT%2520%25C2%25A9%2520FRANCIS%2520AMIAND%2520(2).jpg;https://media.architecturaldigest.com/photos/62d6d05f0e13d40b82bb49cf/master/w_2580%2Cc_limit/G19_4619%2520copy.jpg;https://media.architecturaldigest.com/photos/6377c1d2abb4b34435a704dc/master/w_2580%2Cc_limit/0113-AD-DENI-06%2520copy.jpg;https://media.architecturaldigest.com/photos/6377c1819bc479d85214d271/master/w_2580%2Cc_limit/1213-AD-GARC-05%2520copy.jpg	true	false	5	apartment	7	20	3000	Breakfast;Towels;Fridge	Jean-Batist
+Роскошная двушка в историческом районе Дюссельдорфа	2022-10-01T08:45:40.284Z	https://cf.bstatic.com/xdata/images/hotel/max1280x900/326256908.jpg?k=bdd20df0174cd044f2ec5d975c3586d019b90387a4ea7a2bdca37d42965c1337&o=&hp=1	https://cf.bstatic.com/xdata/images/hotel/max1280x900/326256908.jpg?k=bdd20df0174cd044f2ec5d975c3586d019b90387a4ea7a2bdca37d42965c1337&o=&hp=1;https://cf.bstatic.com/xdata/images/hotel/max1280x900/326256972.jpg?k=890fc5ba61c48d55487cf087ab0b906e8193250a2567eb4649bb865c9b425c23&o=&hp=1;https://cf.bstatic.com/xdata/images/hotel/max1280x900/326257199.jpg?k=0b60c5afecbafcc358a5fa0369e30b0c6d95c2a76ecbe614cb3157d98fcb3b34&o=&hp=1;https://cf.bstatic.com/xdata/images/hotel/max1280x900/326256930.jpg?k=e7d575f736455a64f81144f4b46a70e0061f0a31a15f415fc22e1922a11ec14a&o=&hp=1;https://cf.bstatic.com/xdata/images/hotel/max1280x900/326256937.jpg?k=cd24fd744c417b25a06e5b3b5f5b32f4a5124321c651ff420efd1c997958db3e&o=&hp=1;https://cf.bstatic.com/xdata/images/hotel/max1280x900/326256936.jpg?k=e8643617e246b762df649d5af59685f3b7c447d3e2692974302a0139cd1c77a7&o=&hp=1	true	true	5	apartment	2	2	1000	Breakfast;Towels;Fridge;Air conditioning;Laptop friendly workspace;Baby seat	Hanz
