+++ package.json
@@ -17,7 +17,7 @@
     "ts": "ts-node"
   },
   "devDependencies": {
-    "@types/node": "18.17.17",
+    "@types/node": "^18.17.17",
     "@typescript-eslint/eslint-plugin": "6.7.0",
     "@typescript-eslint/parser": "6.7.0",
     "eslint": "8.49.0",
