+++ src/cli/cli-application.ts
@@ -1,4 +1,4 @@
-import { Command } from './commands/command.interface.js';
+import { Command } from './commands/types.js';
 import { CommandParser } from './command-parser.js';
 
 type CommandCollection = Record<string, Command>;
