+++ src/cli/commands/help.command.ts
@@ -1,4 +1,4 @@
-import { Command } from './command.interface.js';
+import { Command } from './types.js';
 
 export class HelpCommand implements Command {
   public getName(): string {
