+++ src/cli/commands/import.command.ts
@@ -1,5 +1,5 @@
-import { Command } from './command.interface.js';
-import { TSVFileReader } from '../../shared/libs/file-reader/index.js';
+import { Command } from './types.js';
+import { TSVFileReader } from '../../libs/file-reader/index.js';
 
 export class ImportCommand implements Command {
   public getName(): string {
