+++ src/libs/file-reader/types.ts
