+++ src/main.cli.ts
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
-import { CLIApplication, HelpCommand, ImportCommand, VersionCommand } from './cli/index.js';
+import {CLIApplication, HelpCommand, ImportCommand, VersionCommand} from './cli/index.js';
 
 function bootstrap() {
   const cliApplication = new CLIApplication();
