--- src/shared/libs/file-reader/index.ts
@@ -1,2 +1,0 @@
-export * from './file-reader.interface.js';
-export * from './tsv-file-reader.js';
