--- src/shared/types/category.type.ts
@@ -1,3 +1,0 @@
-export type Category = {
-  name: string;
-}
