--- src/shared/types/index.ts
@@ -1,4 +1,0 @@
-export * from './category.type.js';
-export * from './offer-type.enum.js';
-export * from './user.type.js';
-export * from './offer.type.js';
