--- src/shared/types/offer-type.enum.ts
@@ -1,4 +1,0 @@
-export enum OfferType {
-  Buy = 'Buy',
-  Sell = 'Sell',
-}
