--- src/shared/types/offer.type.ts
@@ -1,14 +1,0 @@
-import { OfferType } from './offer-type.enum.js';
-import { Category } from './category.type.js';
-import { User } from './user.type.js';
-
-export type Offer = {
-  title: string;
-  description: string;
-  postDate: Date;
-  image: string;
-  type: OfferType
-  price: number;
-  categories: Category[];
-  user: User;
-}
