--- src/shared/types/user.type.ts
@@ -1,6 +1,0 @@
-export type User = {
-  email: string;
-  avatarPath: string;
-  firstname: string;
-  lastname: string;
-}
