+++ tsconfig.json
@@ -1,8 +1,9 @@
 {
   "compilerOptions": {
+    "baseUrl": "./",
     "target": "ES2022",
     "module": "NodeNext",
-    "moduleResolution": "NodeNext",
+    "moduleResolution": "node",
     "declaration": true,
     "strict": true,
     "noImplicitAny": true,
@@ -21,11 +22,16 @@
     "outDir": "./dist",
     "emitDecoratorMetadata": true,
     "types": [
-      "node",
+      "node"
     ],
     "lib": [
       "ESNext"
-    ]
+    ],
+    "paths": {
+      "@*": [
+        "src/*"
+      ]
+    }
   },
   "include": [
     "src/**/*.ts"
